requirements.txt
fastapi
uvicorn
mysql-connector-python==9.0.0
jinja2
python-dotenv
python-multipart